added browser router
