+++ src/pages/Summary.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import { SummaryResponse } from "../components/SummaryResponse";
+import { RestartButton } from '../components/RestartButton';
+
+
+
+export const Summary = () => {
+
+  return (
+    <>
+    <SummaryResponse />
+    <RestartButton />
+    </>
+  )
+}
